Refactor: usecaseのIF用の置き場所を変更
