+++ internal/usecase/interface/job/job.go
