+++ internal/usecase/testkit/test_kit.go
@@ -5,8 +5,8 @@ import (
 	"context"
 	"testing"
 
-	"boilerplate-go/internal/usecase/tx"
-	mock_tx "boilerplate-go/internal/usecase/tx/mock"
+	"boilerplate-go/internal/usecase/interface/tx"
+	mock_tx "boilerplate-go/internal/usecase/interface/tx/mock"
 	"boilerplate-go/pkg/xerrors"
 
 	"go.uber.org/mock/gomock"
