+++ internal/usecase/user/user_usecase.go
@@ -10,9 +10,9 @@ import (
 	"boilerplate-go/internal/domain/prefecture"
 	"boilerplate-go/internal/domain/user"
 	"boilerplate-go/internal/observability"
+	"boilerplate-go/internal/usecase/interface/tx"
 	"boilerplate-go/internal/usecase/support/paging"
 	"boilerplate-go/internal/usecase/support/search"
-	"boilerplate-go/internal/usecase/tx"
 	"boilerplate-go/pkg/uuid"
 )
