+++ internal/usecase/user/user_usecase_test.go
@@ -9,9 +9,9 @@ import (
 	"boilerplate-go/internal/domain/user"
 	mock_user "boilerplate-go/internal/domain/user/mock"
 	"boilerplate-go/internal/observability"
+	mock_tx "boilerplate-go/internal/usecase/interface/tx/mock"
 	"boilerplate-go/internal/usecase/support/paging"
 	"boilerplate-go/internal/usecase/testkit"
-	mock_tx "boilerplate-go/internal/usecase/tx/mock"
 	"boilerplate-go/pkg/ptr"
 	"boilerplate-go/pkg/uuid"
